added html template refactoring
