+++ src/templates/add_reference.html
@@ -1,0 +1,9 @@
+<form action="/add_reference" method="POST">
+    <label for="author">Author:</label>
+    <input type="text" name="author" id="author"><br>
+    <label for="title">Title:</label>
+    <input type="text" name="title" id="title"><br>
+    <label for="year">Year:</label>
+    <input type="int" name="year" id="year"><br>
+    <input type="submit" value="Add reference">
+</form>
